stop filtering on softlist
why? because the dream of retroarch being a simpler first step to mess
turned sour, and i note softlists sometimes don't even have working
media devices and so on. We will treat softlist as a media device and
spit out all systems into their devices

Unfortuntely, we were using the existence of a softlist also as a way
to denote the difference between a MAME and a MESS system, since we
can't do that anymore, we have to do this.