a naieve implementation of edit distance
also added a rating to each object and got the softlist
name (prefix) out.

the idea here is good, but the number of matches will
be proportional to string length so we'd have to factor
this against the result. what we really want is
to know how similar the strings are irrespective of
string length, so time to try a library