starting to make a variant key in a default emu
my softlist generation works by picking the supported
softlists out of machines, not vice versa. So if we want
to find the regional variations of a machine, we're going
to have to employ pattern matching.

Its unfortunate that means encoding arbitrary rules for
systems which got named differently per region. TODO: cleaner
atternative