Soflist Romdata SUCCESS: regions now printing correctly in output
For the first time the romdatas should be fairly one-click playable

The regional logic sometimes looks a big odd: why is a German c64
game playing with a Swedish C64? Its because there wasn't an exact
match for German, and so the last match for Euro won.

Also the PAL/NTSC check should probably be done first, as it
is we can call PAL/NTSC 'countries' so that we don't have to pass
direct matches all the way down to the standards check

large clearup still needed