fixed corner case of undefined system types
some systems say they work fine, but are clones
of systems that don't work. not sure that's possible,
seems like a mistake in MESS. e.g.: Apple Macintosh XL
being a clone of Apple Lisa

We have the problem that at the point of determining this,
we haven't grabbed the details of the non-working system, so we
don't know its name. The simple fix is just to set system type to
itself, so we'll have a folder Apple Macintosh XL instead of
Apple Lisa.

Really we should go and do a search on the XML, but
looks like none of these exceptions have any games